# internal/adapter/http/api/model/models.gen.go
// Package model provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.13.4 DO NOT EDIT.
package model

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// Account defines model for Account.
type Account struct {
	Domain string `json:"domain"`
	Name   string `json:"name"`
}

// AddressAlias defines model for AddressAlias.
type AddressAlias struct {
	Addresses []EmailAddress `json:"addresses"`
	Alias     EmailAddress   `json:"alias"`
}

// Auth Authentication credentials
type Auth struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// ChangeAccountPassword defines model for ChangeAccountPassword.
type ChangeAccountPassword struct {
	EmailAddress EmailAddress `json:"emailAddress"`
	NewPassword  string       `json:"newPassword"`
}

// Domain Domain of an email address (e.g. @example.com)
type Domain struct {
	// Domain Domínio
	Domain string `json:"domain"`
}

// DomainAlias Alias for an email domain (e.g. @example.net)
type DomainAlias struct {
	Alias  string `json:"alias"`
	Domain string `json:"domain"`
}

// EmailAddress defines model for EmailAddress.
type EmailAddress = string

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code int32 `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// GetTokenResponse JWT Access Token
type GetTokenResponse struct {
	AccessToken string `json:"accessToken"`
}

// NewAccount defines model for NewAccount.
type NewAccount struct {
	Domain   string `json:"domain"`
	Name     string `json:"name"`
	Password string `json:"password"`
}

// CreateAccountJSONRequestBody defines body for CreateAccount for application/json ContentType.
type CreateAccountJSONRequestBody = NewAccount

// UpdateAccountPasswordJSONRequestBody defines body for UpdateAccountPassword for application/json ContentType.
type UpdateAccountPasswordJSONRequestBody = ChangeAccountPassword

// CreateAddressAliasJSONRequestBody defines body for CreateAddressAlias for application/json ContentType.
type CreateAddressAliasJSONRequestBody = AddressAlias

// UpdateAddressAliasJSONRequestBody defines body for UpdateAddressAlias for application/json ContentType.
type UpdateAddressAliasJSONRequestBody = AddressAlias

// CreateDomainJSONRequestBody defines body for CreateDomain for application/json ContentType.
type CreateDomainJSONRequestBody = Domain

// CreateDomainAliasJSONRequestBody defines body for CreateDomainAlias for application/json ContentType.
type CreateDomainAliasJSONRequestBody = DomainAlias

// GetTokenJSONRequestBody defines body for GetToken for application/json ContentType.
type GetTokenJSONRequestBody = Auth